backend/accounting/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
pydantic==2.9.2
python-multipart==0.0.20
httpx==0.27.2
anthropic==0.39.0

# Auth + rate limit + logging
PyJWT==2.10.0
slowapi==0.1.9

# V1.1 §E-2 · 多來源知識庫抽字 + Meili
pymupdf==1.24.14          # PDF (含 OCR 降級)
python-docx==1.1.2        # .docx
python-pptx==1.0.2        # .pptx
openpyxl==3.1.5           # .xlsx / .xls
Pillow==11.0.0            # 圖片 EXIF + 尺寸
meilisearch==0.33.0       # 全文搜尋 client

# 測試(選配)
pytest==8.3.3
mongomock==4.2.0.post1
